Refactor mnemonic text field and improve sign-in functionality
The mnemonic text field in the sign-in dialog has been refactored and the sign-in functionality improved. TextField has been replaced with EditableText in the typeahead_text_field.dart for more control over the input and validation. The sign-in process now supports clipboard pasting for mnemonic word input and dynamically adjusts for varying mnemonic lengths. Validity of passphrase entry is also checked in real-time. The entire sign-in component is now more robust and user-friendly, with improved form validation and feedback mechanisms.