Enhance display rendering and add account dialog functions
This update introduces the ability to handle SVG images in the TokenIcon widget, employing the Flutter SVG and the Cached Network Image packages to provide better rendering and caching possibilities. Additionally, it introduces an account dialog system using AccountDialogCubit for account management functionality. It presents the option to switch between different derived wallets, each displaying their respective balances. Address interaction in widgets is also improved, and some inconsistency issues in UI styles are addressed and updated for better consistency and responsiveness. The settings and logout buttons, as well as the account addition functionality, are also included in the dialog. Overall, these changes not only improve the user interface but also expand account management capabilities.