Update graphic for empty table state and add Keplr integration
This commit replaces the "Nothing here" text with an image for empty table states in `widget/custom_table.dart`. Moreover, it adds integration with the Keplr wallet, which enables users to connect their Keplr wallet and sign transactions. It also updates the `types.dart` file to include the `kiraHub` namespace in all transaction messages.